+++ src/sv/dark/test/SystemDependencyTest.java
@@ -39,7 +39,7 @@ public class SystemDependencyTest implements GameSystem {
     private long executionCount = 0;
 
     @Override
-    public void update(WorldStateFrame state, double deltaTime) {
+    public void update(WorldStateFrame state, float deltaTime) {
         executionCount++;
 
         // Read the value written by SystemExecutionTest
